stm_user_api: changed to support Juno-r0
The channels 0~15 on Juno-r0 seem not permitted to be writen, this patch
also change the start channel of mmap from 0 to 32768.

The STM channel management policy will allocate the channels
automatically from all unused channels.  So if we want to mmap
from a non-zero channel, we have to config the channel range
manually, like below:

/ # mkdir /config/stp-policy/20100000.stm.abc
/ # mkdir /config/stp-policy/20100000.stm.abc/test
/ # echo 32768 65535 > config/stp-policy/20100000.stm.abc/test/channels

Here '32768' is just the channel number from which we want to mmap STM stimulus
space;  'test' is the policy name which we will use to mmap channel
area.

Signed-off-by: Chunyan Zhang <zhang.chunyan@linaro.org>